Order the centres when calling the get_interest_points function
It calculates the position of the circles by taking into account the
relative position of the middle circle with the marker centre.

There is a minor bug when plotting the result, as there is no
coherency between the contours size and the centres size. Anyway, the
algorithm correctly finds the relative position of the circles.